+++ source/GBHitbox.cpp
@@ -13,7 +13,7 @@ using namespace cugl::scene2;
 
 #define FPS 60
 
-bool Hitbox::init(std::shared_ptr<EnemyModel> enemy, Vec2& pos, const Size& size, float scale, int damage, int duration) {
+bool Hitbox::init(std::shared_ptr<EnemyModel> enemy, Vec2& pos, const Size& size, float scale, int damage, float duration) {
     _drawScale = scale;
     setDuration(duration);
     setDamage(damage);
